asdasd

+++ surf spotter dot net core/surf spotter dot net core/Models/HttpProxy.cs
@@ -71,7 +71,7 @@ namespace surf_spotter_dot_net_core.Models
         }
         
 
-        public async Task<Spot> GetOneSpot(long id)
+        public async Task<Spot> GetOneSpot(int id)
         {
             //skal lige fikses - ikke en liste :)
             Spot spot = new Spot();
